generalize some cfg methods, make from_text method generic to use in fcfg
