+++ pyformlang/cfg/cfg.py
@@ -928,64 +928,6 @@ class CFG(Grammar):
             return True
         return False
 
-    def to_text(self) -> str:
-        """
-        Turns the grammar into its string representation. This might lose some\
-         type information and the start_symbol.
-        Returns
-        -------
-        text : str
-            The grammar as a string.
-        """
-        res = []
-        for production in self._productions:
-            res.append(str(production.head) + " -> " +
-                       " ".join([x.to_text() for x in production.body]))
-        return "\n".join(res) + "\n"
-
-    @classmethod
-    def from_text(cls,
-                  text: str,
-                  start_symbol: Optional[Hashable] = Variable("S")) -> "CFG":
-        """
-        Read a context free grammar from a text.
-        The text contains one rule per line.
-        The structure of a production is:
-        head -> body1 | body2 | ... | bodyn
-        where | separates the bodies.
-        A variable (or non terminal) begins by a capital letter.
-        A terminal begins by a non-capital character
-        Terminals and Variables are separated by spaces.
-        An epsilon symbol can be represented by epsilon, $, ε, ϵ or Є.
-        If you want to have a variable name starting with a non-capital \
-        letter or a terminal starting with a capital letter, you can \
-        explicitly give the type of your symbol with "VAR:yourVariableName" \
-        or "TER:yourTerminalName" (with the quotation marks). For example:
-        S -> "TER:John" "VAR:d" a b
-
-        Parameters
-        ----------
-        text : str
-            The text of transform
-        start_symbol : str, optional
-            The start symbol, S by default
-
-        Returns
-        -------
-        cfg : :class:`~pyformlang.cfg.CFG`
-            A context free grammar.
-        """
-        variables = set()
-        productions = set()
-        terminals = set()
-        for line in text.splitlines():
-            line = line.strip()
-            if not line:
-                continue
-            cls._read_line(line, productions, terminals, variables)
-        return cls(variables=variables, terminals=terminals,
-                   productions=productions, start_symbol=start_symbol)
-
     @classmethod
     def _read_line(cls,
                    line: str,
@@ -1017,15 +959,3 @@ class CFG(Grammar):
                     terminals.add(body_ter)
                     body.append(body_ter)
             productions.add(Production(head, body))
-
-    def is_normal_form(self) -> bool:
-        """
-        Tells is the current grammar is in Chomsky Normal Form or not
-
-        Returns
-        -------
-        is_normal_form : bool
-            If the current grammar is in CNF
-        """
-        return all(
-            production.is_normal_form() for production in self._productions)
